e2e tests: just test on chrome
It's kinda pointless to test everything on every single browser. We should just have extra targeted tests for features that break on specific browsers.
