+++ test/system/forum_topic_test.rb
@@ -1,0 +1,126 @@
+require "application_system_test_case"
+
+class ForumTopicsChromeTest < ChromeSystemTestCase
+  context "Forum topics:" do
+    setup do
+      @user = create(:user, created_at: 1.month.ago)
+      @topic = create(:forum_topic, creator: @user, original_post: build(:forum_post, creator: @user, body: "the original post"))
+      @forum_post = create(:forum_post, topic: @topic, creator: @user, body: "the original comment")
+    end
+
+    context "the post's Edit button" do
+      setup do
+        fast_signin @user
+        visit forum_topic_path(@topic)
+      end
+
+      should "reveal the post's edit form" do
+        within "#forum_post_#{@forum_post.id}" do
+          find(".popup-menu-button").click
+          click_link "Edit"
+        end
+
+        assert_visible "#forum_post_#{@forum_post.id} .edit_forum_post"
+        assert_equal @forum_post.body, find("#forum_post_#{@forum_post.id} .edit_forum_post textarea.dtext").value
+      end
+    end
+
+    context "the post's dropdown menu" do
+      setup do
+        stub_clipboard
+        visit forum_topic_path(@topic)
+      end
+
+      should "copy the post's id when clicking Copy ID" do
+        within "#forum_post_#{@forum_post.id}" do
+          find(".popup-menu-button").click
+          click_link "Copy ID"
+        end
+
+        assert_notice "Copied!"
+        assert_clipboard "forum ##{@forum_post.id}"
+      end
+
+      should "copy the post's link when clicking Copy Link" do
+        within "#forum_post_#{@forum_post.id}" do
+          find(".popup-menu-button").click
+          click_link "Copy Link"
+        end
+
+        assert_notice "Copied!"
+        assert_clipboard(/#{Regexp.escape(forum_post_path(@forum_post))}\z/)
+      end
+    end
+
+    context "clicking Reply on a post" do
+      setup do
+        fast_signin @user
+        visit forum_topic_path(@topic)
+      end
+
+      should "open the topic's reply form pre-filled with a quote of the post" do
+        within "#forum_post_#{@forum_post.id}" do
+          click_link "Reply"
+        end
+
+        editor_value = find("#topic-response textarea.dtext", visible: :visible).value
+        assert_includes editor_value, "#{@user.name} said in forum ##{@forum_post.id}:"
+        assert_includes editor_value, @forum_post.body
+      end
+    end
+
+    context "the vote buttons on a bulk update request's post" do
+      setup do
+        @bur = create(:bulk_update_request, user: @user)
+
+        fast_signin @user
+        visit forum_topic_path(@bur.forum_topic)
+      end
+
+      should "vote on the bur's post when clicked" do
+        within "#forum-post-votes-for-#{@bur.forum_post.id}" do
+          find("[title='Vote up']").click
+        end
+
+        assert_notice "Voted"
+        assert_selector "#forum-post-votes-for-#{@bur.forum_post.id} .vote-score-up", text: @user.name
+      end
+    end
+
+    context "as a moderator" do
+      setup do
+        @moderator = create(:moderator_user, created_at: 1.month.ago)
+
+        fast_signin @moderator
+        visit forum_topic_path(@topic)
+      end
+
+      should "ask for confirmation before deleting the post" do
+        within "#forum_post_#{@forum_post.id}" do
+          find(".popup-menu-button").click
+          click_and_accept_confirm(find_link("Delete"), "Are you sure you want to delete this forum post?")
+        end
+
+        assert_selector "#forum_post_#{@forum_post.id}[data-is-deleted='true']"
+      end
+
+      should "ask for confirmation before locking the topic" do
+        click_and_accept_confirm(find("#subnav-lock"), "Are you sure you want to lock this forum topic?")
+
+        assert_selector ".notice", text: "This topic has been locked."
+      end
+
+      should "ask for confirmation before stickying the topic" do
+        click_and_accept_confirm(find("#subnav-sticky"), "Are you sure you want to sticky this forum topic?")
+
+        assert_selector "#subnav-unsticky"
+      end
+
+      should "ask for confirmation before deleting the topic" do
+        click_and_accept_confirm(find("#subnav-delete"), "Are you sure you want to delete this forum topic?")
+
+        assert_selector "h1 .locked-topic", text: "(deleted)"
+      end
+    end
+  end
+end
